Add basic Elm program
After a bit of a dead end trying to build Elm from source (no way I'm
infecting my system with an install of npm to get Elm), it turns out
there is a nice aur package that works perfectly.

My first impressions of Elm after toying around with it for an hour or
so are great. And it got record field access right.

This commit features a basic searchable list of stored entries.