Add native bindings to SJCL
Elm's ports system is far, far to cumbersome to use for something as
simple as two pure functions. I understand the reasons for doing it and
I encourage the principle, but I just want to use this drop-in library
and not worry about asynchronous calls that fracture the entire program.