Do not rely on SJCL implementation detail
When SJCL encrypts something, it returns a string, which is actually an
encoded json object with various fields. Previously I tried to keep that
structure everywhere, because encoding a json object as a string, and
putting that in a string in a json object, feels wrong. But whatever, I
should not rely on this implementation detail. It is simpler not to.

Now saving entries does actually work.