Implement basic persistence of the vault
This is not thread-safe, handlers can race to write. This will only be an issue
for high-volume writes (and when using the threaded runtime), so for a basic
prototype this is fine. It shouldn't be too hard either way to have a worker
thread that writes data and then put things on a queue.